feat: move components to right place
