+++ src/entities/Basket/BasketGoodsItem/BasketGoodsItem.tsx
