+++ src/entities/Basket/interfaces.ts
