+++ src/entities/Card/CardImage/CardImage.tsx
