+++ src/entities/Card/interfaces.ts
