+++ src/entities/Counter/Counter.tsx
