+++ src/entities/Counter/interfaces.ts
