+++ src/entities/Layout/interfaces.ts
