+++ src/entities/Modal/interfaces.ts
