+++ src/entities/StockStatus/StockStatus.tsx
