+++ src/entities/StockStatus/interfaces.ts
