+++ src/entities/index.tsx
@@ -1,9 +1,8 @@
 export { BasketGoodsItem } from './Basket';
-export { AddToCart, Btn, DeleteBtn } from './Btns';
 export { CardDescription, CardImage, FullCard, ShortCard } from './Card';
 export { Counter } from './Counter/Counter';
 export { Header } from './Header/Header';
 export { Layout } from './Layout/Layout';
-export { Logo } from './Logo/Logo';
+
 export { Modal } from './Modal/Modal';
 export { StockStatus } from './StockStatus/StockStatus';
