+++ src/entities/interfaces.ts
@@ -1,8 +1,6 @@
 export * from './Basket/interfaces';
-export * from './Btns/interfaces';
 export * from './Card/interfaces';
 export * from './Counter/interfaces';
 export * from './Layout/interfaces';
-export * from './Logo/interfaces';
 export * from './Modal/interfaces';
 export * from './StockStatus/interfaces';
