+++ src/index.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './app/index.css';
 import App from './app/App';
 
 const root = ReactDOM.createRoot(document.getElementById('root') as HTMLElement);
