+++ src/shared/Btns/DeleteBtn/DeleteBtn.tsx
