+++ src/shared/Btns/interfaces.ts
