+++ src/shared/Logo/interfaces.ts
