--- src/shared/button/sum.test.ts
@@ -1,5 +1,0 @@
-import { sum } from './sum';
-
-test('adds 4 + 5 to equal 9', () => {
-  expect(sum(4, 5)).toBe(9);
-});
