--- src/shared/button/sum.ts
@@ -1,1 +1,0 @@
-export const sum = (a: number, b: number): number => a + b;
