+++ src/shared/index.ts
@@ -1,0 +1,2 @@
+export { AddToCart, Btn, DeleteBtn } from './Btns';
+export { Logo } from './Logo/Logo';
