+++ src/shared/interfaces.ts
@@ -1,0 +1,2 @@
+export * from './Btns/interfaces';
+export * from './Logo/interfaces';
