+++ src/stories/AddToCart.stories.ts
@@ -1,5 +1,5 @@
 import type { Meta } from '@storybook/react';
-import { AddToCart } from '../components';
+import { AddToCart } from 'src/shared';
 
 const meta: Meta<typeof AddToCart> = {
   title: 'UI/Btns/AddToCart',
